improved queries, and tests, still working with config stuff

+++ setuptables.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # setup config file and table names
 config_file="$HOME/.$(basename "${0}").conf"
 table_name1="actions"
@@ -35,11 +37,7 @@ report(){
 
 # define how to connect to the database
 _connect_db(){
-if [[ "${password}" = "" ]] ; then
-mysql -h ${host_name} -u ${user_name} -D ${database} --reconnect
-else
-mysql -h ${host_name} -u ${user_name} -p${password} -D ${database} --reconnect
-fi
+    mysql -h ${host_name} -u ${user_name} -p${password} -D ${database} --reconnect
 }
 
 # read configuration file or create one
@@ -53,15 +51,15 @@ fi
 
 # option for resetting config file
 if [[ "${runtype}" = "reset" ]] ; then
-    report -q -n "Reseting the configuration will clear ${config_file}. Please enter [Y] to confirm: "
-    read reset_response
-    if [[ "${reset_response}" = "Y" || "${reset_response}" = "y" ]] ; then
-        report -d "Clearing ${config_file}."
-        echo -n "" > "${config_file}"
-        runtype="edit"
+ report -q -n "Reseting the configuration will clear ${config_file}. Please enter [Y] to confirm: "
+  read reset_response
+   if [[ "${reset_response}" = "Y" || "${reset_response}" = "y" ]] ; then
+    report -d "Clearing ${config_file}."
+     echo -n "" > "${config_file}"
+      runtype="edit"
     else
-        report -d "Reset aborted. Exiting."
-        exit 0
+       report -d "Reset aborted. Exiting."
+    exit 0
     fi
 fi
 
@@ -98,32 +96,37 @@ read
 
 #Check to see if MySQL server connection is working
 _test_sql(){
-db_errors=$(_connect_db)
+db_errors=$(mysql -h ${host_name} -u ${user_name} -p${password} 2>&1)
 if [[ "${db_errors}" = *"1045"* ]] ; then
 echo "Error connecting with MySQL server. Sorry."
+set db_errors=1
+else
+echo "Connection to MySQL server seems to be ok."
 fi
 }
 
+#run the SQL test
 _test_sql
 
-#Check to see if tables have already been created
+#Check to see if tables have already been created, if not, create them
 _create_tables(){
-table_errors=$(_connect_db << EOF 
-DESCRIBE ${table_name1};
-EOF) 
+table_errors=$(mysql -h ${host_name} -u ${user_name} -p${password} -D ${database} -e "
+DESCRIBE ${table_name1};" 2>&1)
 if [[ "${table_errors}" = *"1146"* ]] ; then
 echo "Tables have not been set up" 
-_connect_db << EOF 
-CREATE TABLE ${table_name1} (action_type varchar(30) NOT NULL, job_id MEDIUMINT(8) NOT NULL AUTO_INCREMENT, technician varchar(128) NOT NULL,  computer_name varchar(128) NOT NULL, time_started DATETIME NOT NULL, time_ended DATETIME NOT NULL, PRIMARY KEY (job_id));
-CREATE TABLE ${table_name2} (job_id MEDIUMINT(8) NOT NULL, timestamp NOT NULL, error_text VARCHAR(65,535), PRIMARY KEY (job_id))
-CREATE TABLE ${table_name3} (digitization_status VARCHAR(50) NOT NULL, item_id VARCHAR(128) NOT NULL, qc_status VARCHAR(50) NULL, object_type VARCHAR(50) NOT NULL, file_path VARCHAR(256) NULL, media_id VARCHAR(20) NULL, title VARCHAR(256) NULL, series VARCHAR(256), PRIMARY KEY (item_id))
-CREATE TABLE ${table_name4} (item_id VARCHAR(128) NOT NULL, relationship_type VARCHAR(128) NOT NULL, PRIMARY KEY (item_id))
-EOF
-echo "Table set up complete!"
-else
-echo "Tables are already set up!"
+mysql -h ${host_name} -u ${user_name} -p${password} -D ${database} -e "CREATE TABLE ${table_name1} (action_type varchar(30) NOT NULL, job_id MEDIUMINT(8) NOT NULL AUTO_INCREMENT, technician varchar(128) NOT NULL,  computer_name varchar(128) NOT NULL, time_started DATETIME NOT NULL, time_ended DATETIME NOT NULL, PRIMARY KEY (job_id));
+CREATE TABLE ${table_name2} (job_id MEDIUMINT(8) NOT NULL, timestamp DATETIME, error_text VARCHAR(8000), PRIMARY KEY (job_id));
+CREATE TABLE ${table_name3} (digitization_status VARCHAR(50) NOT NULL, item_id VARCHAR(128) NOT NULL, qc_status VARCHAR(50) NULL, object_type VARCHAR(50) NOT NULL, file_path VARCHAR(256) NULL, media_id VARCHAR(20) NULL, title VARCHAR(256) NULL, series VARCHAR(256), PRIMARY KEY (item_id));
+CREATE TABLE ${table_name4} (item_id VARCHAR(128) NOT NULL, relationship_type VARCHAR(128) NOT NULL, PRIMARY KEY (item_id));"
+    echo "Table set up complete!"
+elif [[ "${db_errors}" -ne "1" ]] ; then
+    echo "Tables are already set up! B-)"
+else 
+    echo "You've got database problems. :'("
 fi 
 }
+
+_create_tables
 #create a new record when the user starts the digitization (pulls unique ID from vrecord, recording directory, and technician name)
 
 #reads output of vrecord, looks for/parses error messages
